Remove pre Scala 3.0 final hacks, future proof some Scala 3 stuff
