+++ amm/compiler/src/main/scala-3.1.3+/dotty/ammonite/compiler/AmmCompletionExtras.scala
@@ -1,0 +1,12 @@
+package dotty.ammonite.compiler
+
+import dotty.tools.dotc.interactive.Completion
+
+trait AmmCompletionExtras {
+
+  def maybeBackticked(input: String, hasBackTick: Boolean): Completion =
+    Completion.backtickCompletions(Completion(input, "", Nil), hasBackTick)
+  def backtick(completion: Completion): Completion =
+    Completion.backtickCompletions(completion, completion.label.startsWith("`"))
+
+}
